Add files via upload
This is a simple Python implementation of the Rock, Paper, Scissors game using conditional statements. The game allows two players (or a player vs. computer) to select one of the three options: Rock, Paper, or Scissors. The game logic uses a series of if, elif, and else statements to determine the winner based on the following rules: